combine BufferWriteIO with BufferWriterRestoreManager into one class
